requirements.txt
asgiref==3.8.1
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==65.1.0
django-ckeditor==6.7.1
django-cors-headers==4.6.0
django-js-asset==2.2.0
djangorestframework==3.15.2
pillow==11.0.0
sqlparse==0.5.1
